Build the javascript in ES6
