+++ index.html
@@ -5,8 +5,8 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./styles.css" />
-    <title>Document</title>
+    <title>Awesome Books</title>
+    <link rel="stylesheet" href="./index.css" />
 </head>
 
 <body>
@@ -22,13 +22,13 @@
             </nav>
         </section>
         <div class="date-container">
-            <span id="date" class="date">Lorem ipsum dolor sit amet.</span>
+            <span id="date" class="date"></span>
         </div>
         <section id="list-s" class="List-of-books">
             <h1>All awesome books</h1>
             <div id="book-list"></div>
         </section>
-        <section id="add-new-s" class="add-new-section">
+        <section id="add-new-s" class="add-new-section add-new">
             <h2>Add a new book</h2>
             <form>
                 <ul>
@@ -53,11 +53,11 @@
                 <ul>
                     <li>
                         Our email:
-                        <a href="https://mail.google.com/mail/u/0/?tab=rm&ogbl#inbox">abd181103@gmail.com</a
+                        <a href="https://mail.google.com/mail/u/0/?tab=rm&ogbl#inbox">email@gmail.com</a
               >
             </li>
-            <li>Our phone number: +9647501109931</li>
-            <li>Our adress: 90210 Broadway Blvd</li>
+            <li>Our phone number: +251919435423</li>
+            <li>Our address: 90210 Addis Ababa</li>
           </ul>
         </div>
       </section>
@@ -67,6 +67,6 @@
         </footer>
       </section>
     </section>
-    <script type="module" src="./main.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
