+++ modules/time.js
@@ -1,0 +1,6 @@
+const time = () => {
+  const today = new Date();
+  document.getElementById('date').innerHTML = today;
+};
+
+export default time;
